robKalman-package(robKalman)	R Documentation

robKalman – routines for robust Kalman filtering

Description

robKalman provides routines for robust Kalman filtering. Currently, the ACM-filter and the rLS-Filter are provided.

Details

Package: 	 robKalman

Version: 	 0.1

Date: 	 2006-06-10

Depends: 	 R(>= 2.3.0), methods, graphics, startupmsg, dse1, dse2, MASS

Imports: 	 stats

SaveImage: 	 no

LazyLoad: 	 yes

License: 	 GPL (version 2 or later)

URL: 	 http://www.uni-bayreuth.de/departments/math/org/mathe7/robKalman/

Packaged: 	 Sun June 11 10:16:03 2006; Peter

Built: 	 R 2.4.0dev; i386-pc-mingw32; 2006-06-11 10:20:11; windows

Setup

We work in the setup of the time-invariant, linear, Gaussian state space model (ti-l-G-SSM) with p dimensional states x_t and q dimensional observations y_t, with initial condition

x_0 ~ N_p(a,S),

state equation

x_t = F x_{t-1} + v_t, v_t ~ N_p(0,Q), t>=1,

observation equation

y_t = Z x_t + e_t, e_t ~ N_q(0,V), t>=1,

and where all random variable x_0, v_t, e_t are independent.

Classes

yet to be filled

Methods

yet to be filled

Functions



general recursive filters
+recursiveFilter
 -KalmanFilter 
 -rLSFilter
 -ACMfilter

ACMfilter:
+ACMfilt

GM-estimators for AR models
+arGM

utilities:
+Euclidnorm,  
+rcvmvnorm,      
+Huberize,       
+limitS

simulation of AO contaminated state space models       
+simulateState,
+simulateObs,

Acknowledgement

We thank Paul Gilbert for their help in preparing this package.

Start-up-Banner

You may suppress the start-up banner/message completely by setting options("StartupBanner"="off") somewhere before loading this package by library or require in your R-code / R-session.

If option "StartupBanner" is not defined (default) or setting options("StartupBanner"=NULL) or options("StartupBanner"="complete") the complete start-up banner is displayed.

For any other value of option "StartupBanner" (i.e., not in c(NULL,"off","complete")) only the version information is displayed.

The same can be achieved by wrapping the library or require call into either suppressStartupMessages() or onlytypeStartupMessages(.,atypes="version").

Demos

Demos are available — see demo(package="robKalman")

Author(s)

Peter Ruckdeschel (Maintainer) Peter.Ruckdeschel@itwm.fraunhofer.de,
Bernhard Spangl bernhard.spangl@boku.ac.at,

References

Martin, R.D. and Zeh, J.E. (1978): Generalized M-estimates for Autoregression Including Small-sample Efficiency Robustness
Martin, R.D. (1979): Approximate Conditional-mean Type Smoothers and Interpolators.
Martin, R.D. (1980): Robust Estimation of Autoregressive Models.
Martin, R.D. (1981): Robust Methods for Time Series
Martin, R.D. and Thomson, D.J. (1982): Robust-resistent Spectrum Estimation.
Ruckdeschel, P. (2001) Ans"atze zur Robustifizierung des Kalman Filters. Bayreuther Mathematische Schriften, Vol. 64.
Stockinger, N. and Dutter, R. (1987): Robust Time Series Analysis: A Survey.

[Package robKalman version 0.1 Index]