ACMfilt(robKalman)	R Documentation

ACM filter - S-Plus type arguments

Description

This function realizes the ACM filter with a S-Plus type signature.

Usage


ACMfilt(x, gm, s0=0, psi="Hampel", a=2.5, b=a, c=5.0, flag="weights", lagsmo=TRUE)

Arguments

x 	 univarite time series (vector)

gm 	 list as produced by function arGM, with elements ar containing the AR(p) coefficient estimates, sinnov containing innovation scale estiamtes from AR(p) fits of orders 1 through p; Cx containing an estimate of the p by p autocovariance matrix, and mu, the estimated mean of x

s0 	 scale of nominal Gaussian component of additive noise

psi 	 influence function to be used (default: Hampel's psi function, which is the only one available at the moment)

a,b,c 	 tuning constants for Hampel's psi-function

flag 	 character, if "weights", use psi(t)/t to calculate the weights; if "deriv", use psi'(t)

lagsmo 	 logical, if TRUE, lag p-1 smoothing is performed; else filtering from the top of hat X_t is performed

Details

to be filled

Value

a list with elements

filt.ck 	 the classical Kalman filter

filt 	 the ACM filter

st 	 time-dependent scale parameter estimate

r 	 AR-residuals

Author(s)

Bernhard Spangl bernhard.spangl@boku.ac.at,

References

Martin, R.D. and Zeh, J.E. (1978): Generalized M-estimates for Autoregression Including Small-sample Efficiency Robustness
Martin, R.D. (1980): Robust Estimation of Autoregressive Models.
Martin, R.D. and Thomson, D.J. (1982): Robust-resistent Spectrum Estimation.
Stockinger, N. and Dutter, R. (1987): Robust Time Series Analysis: A Survey.

See Also

internalarGM, internalpsi, internalACM, arGM, recFilter

Examples


require(robKalman)

##  AO model:
set.seed(361)
Eps <- as.ts(rnorm(100))
ar2 <- arima.sim(list(ar = c(1, -0.9)), 100, innov = Eps)
Binom <- rbinom(100, 1, 0.1)
Noise <- rnorm(100,sd = 10)
y <- ar2 + as.ts(Binom*Noise)

y.arGM <- arGM(y, 3)
y.ACMfilt <- ACMfilt(y, y.arGM)

plot(y)
lines(y.ACMfilt$filt, col=2)
lines(ar2,col="green")


[Package robKalman version 0.1 Index]