arGM(robKalman)	R Documentation

GM-estimates for AR parameters

Description

This function realizes Martin's GM-estimates for AR parameters.

Usage


arGM(x, order=1, chr=1.5, iterh=maxiter, cbr=5.0, iterb=maxiter, 
     psi2="Tukey", c=4.0, type="Mallows", k=1.5, maxiter=100, 
     tol=1e-08, equal.LS=FALSE, ...) 

Arguments

x 	 univarite time series (vector)

order 	 order of AR(p) process

chr 	 tuning constant for Huber's psi function

iterh 	 number of iterations for IWLS-alogrithm using Huber's psi function

cbr 	 tuning constant for Tukey's psi function

iterb 	 number of iterations for IWLS-alogrithm using Tukey's psi function

psi2 	 influence function to determine the 'size of z_i', either "Ident", "Huber" or "Tukey"

c 	 tuning constant for psi2

type 	 type of GM-estimates, either "Mallows" or "Schweppe"

k 	 tuning constant for centering

maxiter 	 maximal number of iteration

tol 	 tolerance level

equal.LS 	 logical, for testing purpose only

... 	 further parameters to be passed to the functions HuberM or hubers

Details

to be filled

Value

a list with elements

ar 	 parameter estimate

sinnov 	 scale estimate for the innovations from AR(p) fits of orders 1 through p

Cx 	 an estimate of the p x p autocovariance matrix

mu 	 location estimate of x

sx 	 Huber scale estimate

u,v 	 weights for Mallows- or Schweppe-type GM-estimates

w 	 weights from IWLS algorithm

BH 	 consistency constant for σ when using Huber's psi function

BB 	 consistency constant for σ when using Tukey's psi function

niterh 	 number of iterations for IWLS-alogrithm using Huber's psi function

niterb 	 number of iterations for IWLS-alogrithm using Tukey's psi function

niter.testing 	 for testing purposes only

Author(s)

Bernhard Spangl bernhard.spangl@boku.ac.at,

References

Martin, R.D. and Zeh, J.E. (1978): Generalized M-estimates for Autoregression Including Small-sample Efficiency Robustness
Martin, R.D. (1980): Robust Estimation of Autoregressive Models.
Martin, R.D. and Thomson, D.J. (1982): Robust-resistent Spectrum Estimation.
Stockinger, N. and Dutter, R. (1987): Robust Time Series Analysis: A Survey.

See Also

internalarGM, internalpsi

Examples



[Package robKalman version 0.1 Index]