internalarGM(robKalman)	R Documentation

Internal functions of package robKalman — psi functions

Description

These functions are used internally by package robKalman for the ACM filter — psi-functions

Usage


.ARmodel(x, p)
.invCp(p, s, Phi)
.Weights(p, Z, invCp, type, psi2, c)
.startval(y, Z, tol)
.BH(k=1.345) 
.BB(c=4.685)
.weights(r, s, u, v, psi1, ...)
.IWLS(y, Z, phi.ini, s.ini, u, v, psi1, niter, tol, ...)

Arguments

x 	 univarite time series (vector)

p 	 order of AR(p) process

s 	 vector of sx and innovation scale estimates for AR(p-1) models of order 1 to (p-1)

Phi 	 (p-1)x(p-1) matrix of AR(p-1) model parameters

Z 	 AR(p) model matrix

invCp 	 matrix from function .invCp to compute metric

type 	 type of GM-estimates — character: currently: "Mallows" or "Schweppe"

psi1, psi2 	 type of psi function; current possibilities: "Huber", "Tukey", "Hampel", "Ident"

c, k 	 tuning constants

y 	 response vector of AR(p) model

tol 	 tolerance level

r 	 residuals

s 	 innovations scale parameter

u, v 	 weights

... 	 additional arguments (tuning constants) for influence functions

phi.ini 	 initial AR(p) model parameters

s.ini 	 initial innovations scale parameter

niter 	 maximal number of iterations

Details

to be filled

Value

.ARmodel(x, p) returns design matrix Z and response vector y of an AR(p) model (as list with corresponding elements).
.invCp(p, s, Phi) computes the inverse p x p covariance matrix.
.Weights computes weights for Mallows- or Schweppe-type GM-estimates; returns a list with elements u, v.
.startval computes appropriate starting values; returns a list with elements phi, s.
.BH computes appropriate constant to obtain a consistent estimate for σ when using Huber's psi function.
.BB computes appropriate constant to obtain a consistent estimate for σ when using Tukey's psi function.
.weights computes appropriate weights for reweighting — returns a vector.
.IWLS iteratively reweighted least squares algorithm; returns a list with elements phi, s, w, B, niter.

Author(s)

Bernhard Spangl bernhard.spangl@boku.ac.at,

See Also

internalpsi

[Package robKalman version 0.2.1 Index]