internalrLS(robKalman)	R Documentation

Internal functions of package robKalman for the rLS filter

Description

This function is used internally by package robKalman for the rLS filter

Usage


.rLScorrstep(y, x1, S1, Z, V,  rob1=NULL, b, norm=Euclidnorm, ...) 

Arguments

Z 	 observation matrix (see below)

V 	 observation error covariance (see below)

b 	 clipping height b for the rLS filter

norm 	 a function with a numeric vector x as first argument, returning a norm of x - not necessarily, but defaulting to, Euclidean norm; used by rLS filter to determine "too" large corrections

S1 	 prediction error covariance S_{t|t-1} of the classical Kalman filter

rob1 	 not used here; included for compatibility reasons; set to NULL

y 	 observation y_t

x1 	 (rLS Kalman)- predicted state x_{t|t-1}

... 	 not used here; for compatibility with signatures of other "step"-functions

Details

We work in the setup of the time-invariant, linear, Gaussian state space model (ti-l-G-SSM) with p dimensional states x_t and q dimensional observations y_t, with initial condition

x_0 ~ N_p(a,S),

state equation

x_t = F x_{t-1} + v_t, v_t ~ N_p(0,Q), t>=1,

observation equation

y_t = Z x_t + e_t, e_t ~ N_q(0,V), t>=1,

and where all random variable x_0, v_t, e_t are independent.

For notation, let us formulate the classical Kalman filter in this context:

(0) ininitial step

x_{0|0} = a

with error covariance

S_{0|0} = Cov(x_0-x_{0|0}) = S

(1) prediction step

x_{t|t-1} = F x_{t-1|t-1}, t>=1

with error covariance

S_{t|t-1} = Cov(x_t-x_{t|t-1}) = F S_{t-1|t-1} F' + Q

(2) correction step

x_{t|t} = x_{t|t-1} + K_t (y_t - Z x_{t|t-1}), t>=1

for Kalman Gain

K_t = S_{t|t-1} Z' (Z S_{t|t-1} Z' + V )^-

with error covariance

S_{t|t} = Cov(x_t-x_{t|t}) = S_{t|t-1} - K_t Z S_{t|t-1}

Value

.rLScorrstep(y, x1, S1, Z, V, b) calculates rLS-x_{t|t} for arguments b, x1 =x_{t|t-1} (from rLS-past), y =y_{t}, S1 =S_{t|t-1}, Z, and V as above

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de,

References

Ruckdeschel, P. (2001) Ans"atze zur Robustifizierung des Kalman Filters. Bayreuther Mathematische Schriften, Vol. 64.

See Also

internalKalman

[Package robKalman version 0.1 Index]