calibrateRLS(robKalman)	R Documentation

Calibration of clipping height b

Description

calibrates the clipping height b of the rLS-filter in a time-invariant, linear, Gaussian state space model

Usage


rLScalibrateB(Z, S, V, repl = 100000, eff, r, upto=20)

Arguments

Z 	 observation matrix in the (ti-l-G-SSM); see below

S 	 prediction error covariance matrix (of the classical Kalman filter) in the (ti-l-G-SSM); see below

V 	 observation error covariance matrix in the (ti-l-G-SSM); see below

r 	 SO-contamination radius

eff 	 efficiency w.r.t. classical Kalman filter in the ideal model

repl 	 number of replicates used for a LLN-approximation of the expectations needed in this calibration

upto 	 an upper bound to b used in the zero-search of uniroot within rLScalibrateB

Details

We work in the setup of the time-invariant, linear, Gaussian state space model (ti-l-G-SSM) with p dimensional states x_t and q dimensional observations y_t, with initial condition

x_0 ~ N_p(a,S),

state equation

x_t = F x_{t-1} + v_t, v_t ~ N_p(0,Q), t>=1,

observation equation

y_t = Z x_t + e_t, e_t ~ N_q(0,V), t>=1,

and where all random variable x_0, v_t, e_t are independent.

The clipping height b given Z, V, and prediction error covariance S_{t|t-1} (of the classical Kalman filter) is either calibrated to a given efficiency eff in the ideal model or to given (SO)-radius r about the ideal model.

The expectations needed for this calibration are calculated by a LLN approximation with repl replicates;

Value

a clipping height b for the rLS filter

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de,

Examples


require(robKalman)

##Hyper parameter of a lin. time-inv. Gaussian SSM
SS0  <- matrix(0, 2, 2)
F0   <- matrix(c(.7, 0.5, 0.2, 0), 2, 2)
Q0   <- matrix(c(2, 0.5, 0.5, 1), 2, 2)
Z0   <- matrix(c(1, -0.5), 1, 2)
V0i  <- 1

### limiting prediction error covariance
SS <- limitS(S = SS0, F = F0, Q = Q0, Z = Z0, V = V0i)

### calibration b
# by efficiency in the ideal model
# efficiency  =  0.9
(B1 <- rLScalibrateB(eff = 0.9, S = SS, Z = Z0, V = V0i))
# by contamination radius
# r  =  0.1
(B2 <- rLScalibrateB(r = 0.1, S = SS, Z = Z0, V = V0i))

[Package robKalman version 0.1 Index]